added unit tests

+++ src/main/java/com/brighttalk/userrealm/entity/Realm.java
@@ -1,6 +1,5 @@
 package com.brighttalk.userrealm.entity;
 
-import com.fasterxml.jackson.annotation.JsonProperty;
 import lombok.AllArgsConstructor;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
@@ -17,7 +16,7 @@ public class Realm {
 
     @Id
     @GeneratedValue(strategy = GenerationType.AUTO)
-    long id;
+    int id;
 
     @Column(unique = true, length = 250)
     String name;
@@ -27,4 +26,3 @@ public class Realm {
     @Column(name = "\"key\"", length = 32)
     String key;
 }
-//todo: add logging
